backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
qdrant-client
openai
asyncpg
sqlalchemy
alembic
pytest
pytest-asyncio
httpx
pydantic-settings